refactor/#85: parent/mission 경로 변경
