--- app/src/main/java/com/kiero/presentation/parent/schedule/mission/auto/model/MissionUiModel.kt
@@ -1,11 +1,0 @@
-package com.kiero.presentation.parent.schedule.mission.auto.model
-
-import java.time.LocalDate
-
-data class MissionUiModel(
-    val id: Long? = null,
-    val name: String,
-    val reward: Int,
-    val dueAt: LocalDate,
-    val isCompleted: Boolean = false
-)
